fix(viewer): a drifted comment edits against the line it now sits on
A restored comment whose content moved since it was written mounts onto its
new line, but carried its stored line number — so the edit composer announced
"Editing comment on line 40" while sitting on line 47. The card now reports the
line it was actually mounted onto; the persisted anchor is left untouched, so
re-anchoring still resolves from the original line + block text on the next open.

Also breaks the two import cycles in the editor module: the open composer slot
moves to its own module both composers share, the edit composer moves next to
the comment state it reads, and acting on a context action moves next to the
composer that raises it — leaving line classification a pure leaf.
